requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==5.1.2
djangorestframework==3.15.2
httpie==3.2.3
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
Pygments==2.18.0
PySocks==1.7.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
